+++ github-notification-server/src/handlers/health_check.rs
@@ -1,0 +1,5 @@
+use axum::response::Html;
+
+pub async fn health_check_handler() -> Html<&'static str> {
+  Html("<h1>HealthCheck OK</h1>")
+}
