+++ github-notification-server/src/handlers/mod.rs
@@ -1,0 +1,2 @@
+pub mod health_check;
+pub mod notification;
